Added 3 classes; not yet implemented in FileSystemWatcher
These 3 classes, abstract Form, MatchForm, and PrescoutingForm, will be
used to represent the forms as objects, as they come in from the data
collectors. These classes have not yet been implemented in the main file
(FileSystemWatcher), but I think I generalized them enough that they are
ready to be implemented.